requirements.txt
numpy==1.16.2
opencv-contrib-python==3.3.0.10
opencv-python==3.3.0.10
pandas==0.24.2
Pillow==5.4.1
PyAudio==0.2.11
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
